Some refactoring of common classes
