+++ src/main/java/com/airsoft/game/battleroyale/BattleRoyaleService.java
@@ -5,7 +5,7 @@ import com.airsoft.game.battleroyale.api.BattleRoyalePlayerStateDTO;
 import com.airsoft.game.battleroyale.api.GameJoinRequest;
 import com.airsoft.game.battleroyale.model.BattleRoyaleGame;
 import com.airsoft.game.battleroyale.model.BattleRoyalePlayer;
-import com.airsoft.game.geo.LatLongPoint;
+import com.airsoft.game.battleroyale.common.LatLongPoint;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
